+++ src/misc/helpers.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <QDateTime>
+#include <QString>
+
+namespace kemai::helpers {
+
+QString getDurationString(const QDateTime& beginAt, const QDateTime& endAt);
+
+} // namespace kemai::helpers