+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - About dialog [#95](https://github.com/AlexandrePTJ/kemai/issues/95)
+- Recent time sheets history [#86](https://github.com/AlexandrePTJ/kemai/issues/86)
 
 
 ## [0.9.4] - 2023-05-10